Condense the utils
