--- src/utils/default-randomizer.ts
@@ -1,3 +1,0 @@
-export function defaultRandomizer(max: number): number {
-  return Math.floor(Math.random() * Number(max)) + 1
-}
