+++ src/utils/index.ts
@@ -1,7 +1,3 @@
-export * from './find-matches'
-export * from './is-dice-notation'
-export * from './is-randsum-options'
-export * from './make-rolls'
-export * from './patterns'
-export * from './roll-one-factory'
-export * from './default-randomizer'
+export * from './matchers'
+export * from './type-guards'
+export * from './rollers'
