+++ src/utils/matchers.ts
@@ -1,3 +1,5 @@
+import { Match } from 'types'
+
 export const coreNotationPattern = /(?<coreNotationMatch>\d+[Dd]\d+)/
 const dropHighPattern = /(?<dropHighMatch>[Hh]\d*)/
 const dropLowPattern = /(?<dropLowMatch>[Ll]\d*)/
@@ -27,3 +29,18 @@ export const completeRollPattern = new RegExp(
   `${allPatterns.map((pattern) => pattern.source).join('|')}`,
   'g'
 )
+
+export function findMatches(notations: string): Match[] {
+  let m
+  let matches: Match[] = []
+  while ((m = completeRollPattern.exec(notations)) !== null) {
+    if (m.groups !== null && m.groups !== undefined) {
+      for (const key of Object.keys(m.groups)) {
+        if (m.groups[key] !== undefined) {
+          matches = [...matches, { [key]: m.groups[key] }]
+        }
+      }
+    }
+  }
+  return matches
+}
