--- src/utils/roll-one-factory.ts
@@ -1,7 +1,0 @@
-import { Randomizer } from 'types'
-
-export function rollOneFactory(sides: number, randomizer: Randomizer) {
-  return function rollOne() {
-    return randomizer(sides)
-  }
-}
